Update Create DAQ Subsystem.md
Update to constraint 3.

The reason for the need of a minimum 10 analog inputs was a suggestion from Jesse, as we originally thought we would need more than 2 inputs.  The constraint has been updated

The reason that we have decided to use the USB-6215, compared to other products such as the USB-6003, is due to the fact that it has 4 PWM Outputs, which will be needed in other subsystems.  This product also provides excellent resolution which will allow us to get a precise measure compared to real time.  This product was also approved by Jesse when brought to his attention.